requirements.txt
absl-py==0.13.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
efficientnet-pytorch==0.6.3
google-auth==1.31.0
google-auth-oauthlib==0.4.4
grpcio==1.38.0
idna==2.10
kornia==0.5.4
Markdown==3.3.4
munch==2.5.0
numpy==1.20.3
oauthlib==3.1.1
Pillow==8.2.0
pretrainedmodels==0.7.4
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
timm==0.3.2
torch==1.7.1+cu101
torchaudio==0.7.2
torchvision==0.8.2+cu101
tqdm==4.61.1
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
